Add files via upload
It is well established that the forces between nucleons are transmitted by mesons. According to the meson theory, the quantitative explanation of the nuclear forces was extremely tentative and incomplete. But this theory presents a valuable point of view. It is fairly certain now that the nucleons within nuclear matter are in a state made rather different from their free condition by the proximity of other nucleons .Charge independence of nuclear forces demand the existence of neutral meson as amongst the same type of nucleons (P-P) or (N-N). This force demands the same spin and the same orbital angular momentum. The exchange interaction is produced by only a neutral meson. The involving mesons without electric charge that it gives exchange forces between protons and Neutrons. Also therefore maintains charge independence character. It is evident for the nature of the products that neutral mesons decay by both strong and weak interactions. It means that neutral mesons’ constituents are responsible for the electromagnetic interaction. Dramatically neutral mesons play an important role for both electromagnetic and nuclear forces.